augth comps and pages moved around
