+++ client/src/pages/user/ForgotPassword.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import ForgotPassword from './auth/ForgotPassword';
+import ForgotPassword from '../../components/auth/ForgotPassword';
 
 class ForgotPasswordPage extends Component {
   componentDidMount = () => {
