+++ client/src/pages/user/MyAccount.js
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import { Row, Col } from 'react-materialize';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
-import ChangePassword from './auth/ChangePassword';
-import UploadAvatar from '../components/account/UploadAvatar';
+import ChangePassword from '../../components/auth/ChangePassword';
+import UploadAvatar from '../../components/account/UploadAvatar';
 
 class MyAccountPage extends Component {
     componentDidMount = () => {
