+++ client/src/pages/user/ResetPassword.js
@@ -3,7 +3,7 @@ import { Row, Col } from 'react-materialize';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import ChangePassword from './auth/ChangePassword';
+import ChangePassword from '../../components/auth/ChangePassword';
 
 class ForgotPassword extends Component {
   componentDidMount = () => {
